Save changes from any input to ConfigGen
