+++ src/components/header.js
@@ -1,22 +1,8 @@
 import Text from './textbox';
 import TextArea from './textarea';
 import Check from './checkbox';
-import { useState, useEffect} from 'react';
-import {updateConfigInfo} from './ConfigGen';
 
 function GenerateHeader() {
-    const [formValue, setFormValue] = useState({});
-    const handleInput = (event) => {
-        const inputValue = event.target.value;
-        setFormValue(prevValues => ({
-        ...prevValues,
-        [event.target.name]: inputValue
-        }));
-    };
-
-    useEffect(() => {
-        updateConfigInfo(formValue);
-      }, [formValue]);
 
     return(
     <div>
@@ -25,13 +11,13 @@ function GenerateHeader() {
         </div>
         <table>
             <tbody>
-            <Text title="configName" onChange={(e) => handleInput(e)}/>
-            <Text title="type" onChange={(e) => handleInput(e)}/>
-            <Text title="subType" onChange={(e) => handleInput(e)}/>
-            <Text title="client" onChange={(e) => handleInput(e)}/>
-            <Text title="process" onChange={(e) => handleInput(e)}/>
-            <TextArea title="description" onChange={(e) => handleInput(e)}/>
-            <Text title="eventFilters" onChange={(e) => handleInput(e)}/>
+            <Text title="configName"/>
+            <Text title="type"/>
+            <Text title="subType"/>
+            <Text title="client"/>
+            <Text title="process"/>
+            <TextArea title="description"/>
+            <Text title="eventFilters"/>
             </tbody>
         </table>
         <table>
