+++ src/components/number.js
@@ -1,6 +1,19 @@
+import {useState, useEffect} from 'react';
+import {updateConfigInfo} from './ConfigGen';
 var _ = require('lodash');
 
 function Num(props){
+
+  const [formValue, setFormValue] = useState({});
+  const handleInput = (event) => {
+    const inputValue = event.target.value;
+    setFormValue({[event.target.name]: inputValue});
+  };
+
+  useEffect(() => {
+    updateConfigInfo(formValue);
+  }, [formValue]);
+
   var formattedTitle = _.startCase(props.title)
   if(props.title===undefined){var title=""}
   else{var title=props.title}
@@ -8,16 +21,16 @@ function Num(props){
   else{var activity=props.activity}
   if(props.extra===undefined){var extra=""}
   else{var extra=props.extra}
-    return(
-      <tr>
-        <td>
-          <label htmlFor={title + activity + extra}>{formattedTitle}:</label>
-        </td>
-        <td>
-          <input type="number" id={title + activity + extra} name={title + activity + extra}/>
-        </td>
-      </tr>
-    )
-  }
+  return(
+    <tr>
+      <td>
+        <label htmlFor={title + activity + extra}>{formattedTitle}:</label>
+      </td>
+      <td>
+        <input type="number" id={title + activity + extra} name={title + activity + extra}  onChange={(e) => handleInput(e)}/>
+      </td>
+    </tr>
+  )
+}
 
-  export default Num;
+export default Num;
