+++ src/components/textbox.js
@@ -1,8 +1,20 @@
-import {useState} from 'react';
+import {useState, useEffect} from 'react';
+import {updateConfigInfo} from './ConfigGen';
 var _ = require('lodash');
 
-function Text(props){  
-  // const [formValues, setFormValues] = useState({});
+function Text(props){
+
+  const [formValue, setFormValue] = useState({});
+  const handleInput = (event) => {
+    const inputValue = event.target.value;
+    setFormValue({[event.target.name]: inputValue});
+  };
+
+  useEffect(() => {
+    updateConfigInfo(formValue);
+  }, [formValue]);
+
+  var formattedTitle = _.startCase(props.title)
   if(props.title===undefined){var title=""}
   else{var title=props.title}
   if(props.activity===undefined){var activity=""}
@@ -10,17 +22,16 @@ function Text(props){
   if(props.extra===undefined){var extra=""}
   else{var extra=props.extra}
   
-  var formattedTitle = _.startCase(props.title)
-    return(
-      <tr>
-        <td>
-          <label htmlFor={title + activity + extra}>{formattedTitle}:</label>
-        </td>
-        <td>
-          <input type="text" id={title + activity + extra} name={title + activity + extra} onChange={props.onChange}/>
-        </td>
-      </tr>
-    )
-  }
+  return(
+    <tr>
+      <td>
+        <label htmlFor={title + activity + extra}>{formattedTitle}:</label>
+      </td>
+      <td>
+        <input type="text" id={title + activity + extra} name={title + activity + extra}  onChange={(e) => handleInput(e)}/>
+      </td>
+    </tr>
+  )
+}
 
 export default Text;
